fix path0
